+++ kernel-stm32l4/build.rs
@@ -8,7 +8,11 @@ fn main() {
     let out = &PathBuf::from(env::var_os("OUT_DIR").unwrap());
     File::create(out.join("memory.x"))
         .unwrap()
-        .write_all(include_bytes!("../memory.x"))
+        .write_all(include_bytes!("memory.x"))
+        .unwrap();
+    File::create(out.join("mwatch.x"))
+        .unwrap()
+        .write_all(include_bytes!("mwatch.x"))
         .unwrap();
     println!("cargo:rustc-link-search={}", out.display());